src/officehours/captions/requirements.txt
vtt_to_srt3
pysrt